Change how we parse ~properties
In Neo4j the properties of a node or relationship were at the level as
the id, and type keys.

In Neptune these properties are stored in a hash under a "~properties"
key. We hoist these values to be at the same level as Neo4j so our
query results are formatted the same way